+++ relay-conventions/src/consts.rs
@@ -46,6 +46,7 @@ convention_attributes!(
     USER_GEO_COUNTRY_CODE => "user.geo.country_code",
     USER_GEO_REGION => "user.geo.region",
     USER_GEO_SUBDIVISION => "user.geo.subdivision",
+    USER_AGENT_ORIGINAL => "user_agent.original",
 );
 
 /// Attributes which are in use by Relay but are not yet defined in the Sentry conventions.